% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-input.r
\name{fake.some.portfolio.data}
\alias{fake.some.portfolio.data}
\title{Generate fake input data for
personal investment portfolio
finance}
\usage{
fake.some.portfolio.data(
  start.date = "2010-01-01",
  end.date = Sys.Date(),
  n = 20,
  tickers = c("AAPL", "NFLX"),
  pct.bought = 0.5,
  purchase.ceiling = 5000
)
}
\arguments{
\item{start.date}{A start date in a format understood by \code{as.Date} to be the earliest date in the faked data.}

\item{end.date}{An end date in a format understood by \code{as.Date} to be the latest date in the faked data.}

\item{n}{The number of records (rows) to generate.}

\item{tickers}{The stock \code{ticker} strings of the companies desired in the generated dataset.}

\item{pct.bought}{The percentage of random data per ticker to be labeled "bought" (as opposed to "sold").}
}
\value{
Returns a simulated data.table
}
\description{
Generate fake input data for
personal investment portfolio
finance
}
