% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-input.r
\name{generate.fake.dates}
\alias{generate.fake.dates}
\title{Generate a list of random dates within a period of time.}
\usage{
generate.fake.dates(
  start.date = "2010-01-01",
  end.date = Sys.Date(),
  n.dates = 100
)
}
\arguments{
\item{start.date}{A start date in a format understood by \code{as.Date} to be the earliest date in the faked data.}

\item{end.date}{An end date in a format understood by \code{as.Date} to be the latest date in the faked data.}

\item{n.dates}{The number of dates to generate.}
}
\value{
Returns a list of dates generated.
}
\description{
Generate a list of random dates within a period of time.
}
